Fix tests, remove Techs base_tech validation to avoid false negatives when using data tables
