+++ src/calliope/preprocess/data_tables.py
@@ -15,12 +15,7 @@ from typing_extensions import NotRequired, TypedDict
 from calliope import exceptions
 from calliope.attrdict import AttrDict
 from calliope.io import load_config
-from calliope.util.schema import (
-    DATA_TABLE_SCHEMA,
-    MODEL_SCHEMA,
-    extract_from_schema,
-    validate_dict,
-)
+from calliope.util.schema import MODEL_SCHEMA, extract_from_schema
 from calliope.util.tools import listify, relative_path
 
 LOGGER = logging.getLogger(__name__)
@@ -59,7 +54,7 @@ class DataTable:
 
         Args:
             table_name (str): name of the data table.
-            data_table (DataTableDict): Data table definition dictionary.
+            data_table (DataTableDict): validated data table definition dictionary.
             data_table_dfs (dict[str, pd.DataFrame] | None, optional):
                 If given, a dictionary mapping table names in `data_table` to in-memory pandas DataFrames.
                 Defaults to None.
@@ -68,7 +63,6 @@ class DataTable:
                 If None, relative data table filepaths will be considered relative to the current working directory.
                 Defaults to None.
         """
-        validate_dict(data_table, DATA_TABLE_SCHEMA, "data table")
         self.input = data_table
         self.dfs = data_table_dfs if data_table_dfs is not None else dict()
         self.model_definition_path = model_definition_path
